Port Hermes migration across master module splits
