+++ packages/agent-cli/src/Agent/CLI/AgentSessions.hs
@@ -23,6 +23,7 @@ module Agent.CLI.AgentSessions
 
 import Agent.CLI.Options (ApprovalPolicy(..))
 import Agent.CLI.ManagedTurn (ManagedTurnRequest)
+import Agent.CLI.AgentSessions.Render (renderAgentSession)
 import Agent.CLI.Error (formatException)
 import Agent.Process
     ( terminateProcessGroupWith
@@ -30,10 +31,8 @@ import Agent.Process
     )
 import Agent.CLI.Session
     ( SessionCreate(..)
-    , SessionActivity(..)
     , SessionHandle(..)
     , SessionMeta(..)
-    , SessionTurn(..)
     , SessionTurnPage(..)
     , createSession
     , loadSessionActivity
@@ -57,11 +56,10 @@ import Agent.OsPath (fromText, unsafeToFilePath)
 import Agent.Dialect
     ( DialectId
     , dialectIdForModel
-    , dialectSlug
     )
-import Agent.Provider (Provider, providerSlug)
+import Agent.Provider (Provider)
 import Agent.Json.Decode (optionalKey)
-import Agent.Json.Decode qualified as Hermes
+import qualified Agent.Json.Decode as Hermes
 import Agent.ToolDSL (PropertySchema(..), PropertyType(..))
 import Agent.ToolDispatch (typedTool)
 import Agent.Tools.Types
@@ -987,63 +985,3 @@ statusAfterLaunch env sessionId launchResult
 renderSessionLaunch :: Text -> Text -> Text
 renderSessionLaunch sessionId status =
     "Session: " <> sessionId <> "\nStatus: " <> status
-
-renderAgentSession
-    :: SessionMeta
-    -> Text
-    -> Maybe SessionActivity
-    -> [SessionTurn]
-    -> Text
-renderAgentSession meta status activity turns =
-    Text.intercalate "\n" $
-        [ "Session"
-        , "  ID: " <> meta.metaId
-        , "  Status: " <> status
-        , "  Title: " <> meta.metaTitle
-        , "  Provider: " <> providerSlug meta.metaProvider
-        , "  Connection: " <> meta.metaConnection
-        , "  Model: " <> meta.metaModel
-        , "  Dialect: " <> dialectSlug meta.metaDialect
-        , "  Reasoning effort: " <> meta.metaEffort
-        , "  Working directory: " <> Text.pack (unsafeToFilePath meta.metaCwd)
-        , "  Created at: " <> Text.pack (show meta.metaCreatedAt)
-        , "  Updated at: " <> Text.pack (show meta.metaUpdatedAt)
-        ]
-            <> maybe [] renderActivity activity
-            <> ["", "Recent turns: " <> Text.pack (show (length turns))]
-            <> case turns of
-                [] -> ["  (none)"]
-                _ -> [""] <> intercalateBlank
-                    (zipWith renderSessionTurn [1 :: Int ..] turns)
-
-renderActivity :: SessionActivity -> [Text]
-renderActivity activity =
-    [ ""
-    , "Current activity"
-    , "  Kind: " <> activity.activityKind
-    , "  Message: " <> activity.activityMessage
-    ]
-        <> maybe []
-            (\retryAt -> ["  Retry at: " <> Text.pack (show retryAt)])
-            activity.activityRetryAt
-        <> ["  Updated at: " <> Text.pack (show activity.activityUpdatedAt)]
-
-renderSessionTurn :: Int -> SessionTurn -> [Text]
-renderSessionTurn index turn =
-    [ "Turn " <> Text.pack (show index)
-    , "At: " <> Text.pack (show turn.turnAt)
-    , "User:"
-    , indentText turn.turnUserText
-    ]
-        <> maybe [] (\text -> ["Assistant:", indentText text])
-            turn.turnAssistantText
-        <> maybe [] (\text -> ["Error:", indentText text])
-            turn.turnError
-
-intercalateBlank :: [[Text]] -> [Text]
-intercalateBlank = \case
-    [] -> []
-    first : rest -> first <> concatMap ("" :) rest
-
-indentText :: Text -> Text
-indentText = Text.intercalate "\n" . map ("  " <>) . Text.lines
